feat(editor-content): fix error with sonar #29872
